# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Alyra Solutions - Multi-tenant Nail Salon SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /salons/{salonId}/users:
    get:
      operationId: listSalonUsers
      tags: [users]
      summary: List users in salon
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: query
          name: role
          schema:
            type: string
            enum: [salon_owner, technician, client]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"

    post:
      operationId: createSalonUser
      tags: [users]
      summary: Create user in salon
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /salons/{salonId}/users/{userId}:
    get:
      operationId: getSalonUser
      tags: [users]
      summary: Get user details
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    put:
      operationId: updateSalonUser
      tags: [users]
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteSalonUser
      tags: [users]
      summary: Delete user from salon
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Technicians
  /salons/{salonId}/technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List technicians in salon
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: query
          name: available
          schema:
            type: boolean
      responses:
        "200":
          description: List of technicians
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianListResponse"

    post:
      operationId: createTechnician
      tags: [technicians]
      summary: Create technician profile
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTechnicianRequest"
      responses:
        "201":
          description: Technician created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /salons/{salonId}/technicians/{technicianId}:
    get:
      operationId: getTechnician
      tags: [technicians]
      summary: Get technician details
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/technicianId"
      responses:
        "200":
          description: Technician details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"

    put:
      operationId: updateTechnician
      tags: [technicians]
      summary: Update technician
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/technicianId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTechnicianRequest"
      responses:
        "200":
          description: Technician updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /salons/{salonId}/technicians/{technicianId}/schedule:
    get:
      operationId: getTechnicianSchedule
      tags: [technicians]
      summary: Get technician schedule/availability
      parameters:
        - $ref: "#/components/parameters/salonId"
        - $ref: "#/components/parameters/technicianId"
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Technician schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianSchedule"

  # Services
...
  /salons/{salonId}/waitlist:
    get:
      operationId: getWaitlist
      tags: [waitlist]
      summary: Get current waitlist
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: query
          name: status
          schema:
            type: string
            enum: [waiting, called, serving, completed, left]
      responses:
        "200":
          description: Current waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistResponse"

    post:
      operationId: joinWaitlist
      tags: [waitlist]
      summary: Add walk-in to waitlist
      parameters:
        - $ref: "#/components/parameters/salonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinWaitlistRequest"
      responses:
        "201":
          description: Added to waitlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"
  /salons/{salonId}/waitlist/{entryId}:
    get:
      operationId: getWaitlistEntry
      tags: [waitlist]
      summary: Get waitlist entry status
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: path
          name: entryId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Waitlist entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"

    put:
      operationId: updateWaitlistEntry
      tags: [waitlist]
      summary: Update waitlist entry status
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: path
          name: entryId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWaitlistEntryRequest"
      responses:
        "200":
          description: Waitlist entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"
  /salons/{salonId}/waitlist/qr-code:
    get:
      operationId: getWaitlistQrCode
      tags: [waitlist]
      summary: Get QR code for walk-in check-in
      parameters:
        - $ref: "#/components/parameters/salonId"
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeResponse"
  /salons/{salonId}/waitlist/wait-time:
    get:
      operationId: getEstimatedWaitTime
      tags: [waitlist]
      summary: Get estimated wait time
      parameters:
        - $ref: "#/components/parameters/salonId"
        - in: query
          name: serviceIds
          schema:
            type: string
      responses:
        "200":
          description: Estimated wait time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitTimeResponse"

  # Nail Colors
...
components:
  parameters:
    salonId:
      in: path
      name: salonId
      required: true
      schema:
        type: integer
    userId:
      in: path
      name: userId
      required: true
      schema:
        type: integer
    technicianId:
      in: path
      name: technicianId
      required: true
      schema:
        type: integer
...
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages

    # Auth
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [platform_admin, salon_owner, technician, client]
        salonId:
          type: integer
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [salon_owner, technician, client]
      required:
        - email
        - password
        - firstName
        - lastName
        - role
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # Salons
...
    DayHours:
      type: object
      properties:
        open:
          type: boolean
        openTime:
          type: string
          nullable: true
        closeTime:
          type: string
          nullable: true
      required:
        - open
...
    Technician:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        salonId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        specialties:
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        isAvailable:
          type: boolean
        rating:
          type: number
          nullable: true
        totalReviews:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - salonId
        - firstName
        - lastName
        - email
        - specialties
        - isActive
        - isAvailable
        - totalReviews
        - createdAt
    CreateTechnicianRequest:
      type: object
      properties:
        userId:
          type: integer
        specialties:
          type: array
          items:
            type: string
        bio:
          type: string
      required:
        - userId
        - specialties
    UpdateTechnicianRequest:
      type: object
      properties:
        specialties:
          type: array
          items:
            type: string
        bio:
          type: string
        isActive:
          type: boolean
        isAvailable:
          type: boolean
    TechnicianListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Technician"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    TechnicianSchedule:
      type: object
      properties:
        technicianId:
          type: integer
        date:
          type: string
          format: date
        workHours:
          $ref: "#/components/schemas/DayHours"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/AppointmentSlot"
        blockedSlots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required:
        - technicianId
        - date
        - appointments
        - blockedSlots
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
      required:
        - startTime
        - endTime
    AppointmentSlot:
      type: object
      properties:
        appointmentId:
          type: integer
        clientName:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        services:
          type: array
          items:
            type: string
        status:
          type: string
      required:
        - appointmentId
        - clientName
        - startTime
        - endTime
        - services
        - status

    # Services
...
    WaitlistEntry:
      type: object
      properties:
        id:
          type: integer
        salonId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
          nullable: true
        partySize:
          type: integer
        serviceIds:
          type: array
          items:
            type: integer
        preferredTechnicianId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [waiting, called, serving, completed, left]
        position:
          type: integer
        estimatedWaitMinutes:
          type: integer
        notes:
          type: string
          nullable: true
        checkInCode:
          type: string
        joinedAt:
          type: string
          format: date-time
        calledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        clientId:
          type: integer
          nullable: true
      required:
        - id
        - salonId
        - clientName
        - partySize
        - serviceIds
        - status
        - position
        - estimatedWaitMinutes
        - checkInCode
        - joinedAt
    JoinWaitlistRequest:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
        partySize:
          type: integer
          default: 1
        serviceIds:
          type: array
          items:
            type: integer
        preferredTechnicianId:
          type: integer
        notes:
          type: string
        clientId:
          type: integer
      required:
        - clientName
        - partySize
        - serviceIds
    UpdateWaitlistEntryRequest:
      type: object
      properties:
        status:
          type: string
          enum: [waiting, called, serving, completed, left]
        preferredTechnicianId:
          type: integer
    WaitlistResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/WaitlistEntry"
        totalWaiting:
          type: integer
        avgWaitTime:
          type: integer
      required:
        - entries
        - totalWaiting
        - avgWaitTime
    QrCodeResponse:
      type: object
      properties:
        qrCodeUrl:
          type: string
        checkInUrl:
          type: string
        salonId:
          type: integer
      required:
        - qrCodeUrl
        - checkInUrl
        - salonId
    WaitTimeResponse:
      type: object
      properties:
        estimatedMinutes:
          type: integer
        currentQueueLength:
          type: integer
        availableTechnicians:
          type: integer
      required:
        - estimatedMinutes
        - currentQueueLength
        - availableTechnicians

    # Nail Colors
...